Tests/LibWeb/Text/expected/css/transition-legacy-event-listener.txt

legacy transitionend event count: 1
legacy transitionend event type: webkitTransitionEnd
legacy transitionend property name: opacity
